removal of iframe js
there really is no need to have an iframe specific feature for the modal. It could potentially be kept as an add-on...but for out of the box, if anyone wanted this corner-case feature, they could also implement it as they see fit into one of the regular template polyfill scripts.